Fix warning

+++ src/codec_aom.c
@@ -158,7 +158,7 @@ static avifBool aomCodecGetNextImage(avifCodec * codec, avifImage * image)
             nclx.colourPrimaries = (uint16_t)codec->internal->image->cp;
             nclx.transferCharacteristics = (uint16_t)codec->internal->image->tc;
             nclx.matrixCoefficients = (uint16_t)codec->internal->image->mc;
-            nclx.fullRangeFlag = image->yuvRange;
+            nclx.fullRangeFlag = (uint8_t)image->yuvRange;
             avifImageSetProfileNCLX(image, &nclx);
         }
